Implement divisible_by_7, update parameterized tests
Finalize the `divisible_by_7` implementation based on the first
algorithm seen in the "Why 7 is Weird" video. This implementation was
chosen as it avoids unsigned integer underflow.

The parameterized tests for `divisible_by_7` were updated as the cases
listed previously appear explicitly in the implementation. The new
negative cases test boundary cases (for the current implementation), the
new positive cases are the first "non-trivial" example as well as the
examples from the "Why 7 is Weird" video.