implement `divisible_by_1`, update parameterized tests
Finalize the `divisible_by_1` implementation based on the algorithm seen
in the "Why 7 is Weird" video.

The parameterized test for `divisible_by_1` was created based on values
obtained from the property test due to partial implementations of the
function. The options were limited to 5 based on the number of options
for pre-existing parameterized tests