Start `divisible_by_3` development
Start the implementation of the `divisible_by_3` function in a manner
similar to the `divisible_by_2` function. This includes a **failing**
property test for numbers that should be divisible by 3. This property
test is, again, accompanied by a corrresponding parameterized tests as a
regression for the failing test case detected by the property test.