Set up minimal CI
Based on the GitHub-provided workflow template for Rust as well as the
resources of the (unofficial) actions-rs GitHub organization [1].

The build downloads the appropriate Rust toolchain, attempts to build
the library, checks the formatting of the Rust code, and runs the tests.
Cargo resources and build outputs are cached to improve build speeds -
this is based on the example for Rust from the actions/cache docs.

All Actions used in the workflow are intentionally pinned to specific
version in order to (in a reasonably easy manner) have consistent
behavior across builds.

--
1. https://github.com/orgs/actions-rs