Start `divisible_by_5` development
Start the implementation of the `divisible_by_5` the same way as that of
`divisible_by_7`. The TDD approach will be slightly different this time
around. Rather than starting with one failing test case or starting with
one failing property test, this time we start with a collection of
failing tests for common as well as curated "edge cases" (positive edge
cases for divisibility by 5 are hard to come by...).