Start `divisible_by_8` development
Start the implementation of the `divisible_by_8` this time starting with
property tests, like `divisible_by_4`, but also leveraging the
proptest-regressions directory to re-run regression cases that occurred
in the past.

Per the proptest documentation,

> [What] we should do is copy the failing case to a traditional unit
> test since it has exposed a bug not similar to what we've tested in
> the past.

elaborating on which, this is necessary in case the property test is
removed or changed, in which case the regression seed likely results in
a different test case.