Implement `digit_sum`
Implement the `digit_sum` function, the reader is spared from having to
view the TDD process on this one :)

The result is an implementation closely matching what's seen in `lib.rs`
with a few tests. First, the parameterized test is expanded with some
additional cases capturing specific scenarios. Second, there's one
property tests - coming up with properties for this functions is pretty
hard, contributions are welcome - that ensures computing digit sums
always results in a smaller number (this is useful because the digit sum
is used for recursive functions in `lib.rs`).