Start `alternating_digit_sum` implementation
Start the implementation of a (new) utility function called
`alternating_digit_sum` that will be extracted from the
`divisible_by_11` function.

The function is currently lacking an implementation and has one
corresponding, failing, parameterized unit test. Like with `last_digit`
and `digit_sum` the reader will be spared the steps of the TDD approach
of the implementation process.