Set up benchmarking
Create a basic benchmarking setup using _bencher_ (new dev-dependency).
The added benchmarks compare safe against unsafe versions of all utility
functions - unsafe meaning that `unwrap`s are done unchecked (which is
OK because we know certain properties about the `Option`s that `unwrap`
is called on due to the `Option`s originating from `u64::to_string`).

Running the benchmarks shows a slight improvement in performance (to be
expected due to `unwrap_unchecked` lacking a check for `None` compared
to `unwrap`). Example results from my machine:

    $ cargo bench
    [...]
    test alternating_digit_sum_safe   ... bench:      41 ns/iter (+/- 1)
    test alternating_digit_sum_unsafe ... bench:      40 ns/iter (+/- 2)
    test digit_sum_safe               ... bench:      41 ns/iter (+/- 2)
    test digit_sum_unsafe             ... bench:      39 ns/iter (+/- 2)
    test last_digit_safe              ... bench:      33 ns/iter (+/- 0)
    test last_digit_unsafe            ... bench:      31 ns/iter (+/- 0)

Given these results, the implementations of these functions (in `src/`)
have been updated to be the unsafe variant as well.

The CI configuration is also updated to include a job to run all the
benchmarks. The goal of this, for now, is primarily to ensure they can
be run without problems (aka, correctness). Note that the benchmarks are
ran after the tests, this is because I see no point in benchmarking a
function that doesn't do what it's supposed to do.