Add Doc-tests to public functions
Update the documentation for public functions to include at least one
example each. Additionally, the in-code documentation now uses the Rust
idiomatic `///` for rustdoc. Imports are being omitted from examples for
brevity.

As Doc-tests are ran automatically by `cargo test`, no changes to the CI
are needed to run them continuously.

On top of all that, this also adds basic "front-page" documentation for
this library. The copy is based on the description in the README.md, but
exclude details not relevant to potential users of the library (such as
the use of the TDD methodology).