Bump Kani to v0.20.0, Kani proofs for utils.rs
Upgrade to Kani[1] v0.20.0[2] to make it possible to verify non-trivial
functions in this project. This was previously not possible due to a bug
in the CMBC[3] version used by Kani. Accordingly, upgrade the version of
the `model-checking/kani-github-action` Action[4] to the latest version
as well.

With this new Kani version, create `kani::proof`s for all functions in
the utils.rs file. Each proof is based on a property test for the same
function. Explicit unwinding[5] limits are used to prevent proof
non-determination. The exact values are minimums acquired by starting at
an `unwind` value of 1 and going up until the proofs succeed.

The proofs for both `alternating_digit_sum` and `digit_sum` are defined
in terms of `u16`s instead of `u64`s (or `u32`s) to bound the proof size
to a reasonable amount, making it feasible to run the proofs. With `u32`
I was able to complete the proofs for both of these in ~45 seconds (with
an `unwind` value of 11). This isn't infeasible, but the `u16`-based
proofs provide sufficient guarantees on their correctness.

The proof for `last_digit` influenced the property test in that it was
rewritten to be more concise (matching any if the first 10 digits is
equivalent to being smaller than 10).

--
1. https://github.com/model-checking/kani
2. https://github.com/model-checking/kani/releases/tag/kani-0.20.0
3. https://github.com/diffblue/cbmc
4. https://github.com/marketplace/actions/kani-rust-verifier
5. https://model-checking.github.io/kani/tutorial-loop-unwinding.html