Kani proofs with limited unwinding
Create Kani proofs for all (previously unchecked) divisibility checks
that were found to succeed and terminate when using the low unwinding
(and the `--no-unwinding-checks` option, to prevent errors as a result).

Lowering the unwinding value reduces proof bounds, but it was found that
coverage (using `--visualize`) is not bad for all of the added proofs.
Only 6 and 12 have lower than expected coverage, seemingly due to the &&
over two other divisibility checks.

One upside of the limited unwinding value is that there's no bound
necessary on the input value, so `u64` can be used for all these proofs.

This only excludes the divisibility check for 7, which fails even when
using `#[kani::unwind(1)]`.