Use Kani Verifier (#1)
* Kani Verifier proof of concept

Add Kani Verifier [1,2] as a dev-dependency and the corresponding GitHub
Action [3] to the CI workflow. Initialize usage of the Kani Verifier by
adding two simple `kani::proof`s for the (trivial) `divisible_by_0` and
`divisible_by_1` functions.

* Kani proofs for utils.rs

Create Kani proofs for all functions in the utils.rs file. Each proof is
based on a property test for the same function. Explicit unwinding [4]
limits are used to prevent proof non-determination. The exact values are
minimums acquired by starting at an `unwind` value of 1 and going up
until the proofs succeed.

The proofs for both `alternating_digit_sum` and `digit_sum` are defined
in terms of `u16`s instead of `u64`s (or `u32`s) to bound the proof size
to a reasonable amount; making it feasible to run the proofs. With `u32`
I was able to complete the proofs for both of these in ~45 seconds (with
an `unwind` value of 11). This isn't infeasible, but I think the
`u16`-based proofs provide sufficient guarantees on their correctness.

The proof for `last_digit` actually influenced the property test in that
it was rewritten to be more concise (matching any if the first 10 digits
is equivalent to being smaller than 10).

* Kani proofs for "simplest" divisibility checks

Create Kani proofs for all divisibility checks that were found to have a
proof harness that terminates in a "reasonable" amount of time (not more
than 20 seconds). The present bounds (in terms integer size) were chosen
to optimize performance. For some of the harnesses, proof using `u16`,
`u32` or `u64` would terminate in 50+ seconds. For others the proof
would not terminate even after 5 minutes.

This covers divisibility by 2, 4, 5, 8, and 10.

* Kani proofs for divisibility checks with limited unwinding

Create Kani proofs for all divisibility checks that were found to
succeed and terminate when using the low unwinding (and the option
`--no-unwinding-checks`, to prevent errors as a result).

Lowering the unwinding value reduces proof bounds. Yet, it was found
that coverage (using `--visualize`) is not bad for all of the added
proofs. Only 6 and 12 have lower than expected coverage, seemingly due
to the `&&` over two other divisibility checks.

One upside of the limited unwinding value is that there's no bound
necessary on the input value size, so `u64` can be used for all these
proofs.

This covers divisibility by 3, 6, 9, 11, and 12.

* Kani proof for `divisible_by_7` + bugfix

Create a proof (with limited unwinding as well) for the divisibility by
7 function - which uncovered a bug!

The bug is that the divisibility check returned the wrong answer for the
value 0. This has been patched with an accompanying unit test to prevent
regression (in a way that is quicker than running Kani).

--
1. https://crates.io/crates/kani-verifier
2. https://github.com/model-checking/kani
3. https://github.com/marketplace/actions/kani-rust-verifier
4. https://model-checking.github.io/kani/tutorial-loop-unwinding.html