Use unsafe `unwrap_unchecked` in `divisible_by_7`
Refactor the `divisible_by_7` function to perform unchecked unwrapping
of `Option`s and `Result`s for slightly improved performance. Tests and
Kani proofs should provide sufficient guarantees that this is safe.

In particular:
- L125: This unwrapping should be safe since `n_as_str` is the string
  representation of a number greater than 50. Such a number will always
  be more than 2 characters long, i.e. `n_len >= 2`, and hence the
  range `0..n_len-1` will always exist (ranging from `0..1` to `0..64`)
  and because `n_len = n_as_str.len()` never exceed the string length.
- L126: Following from the above, `rest_as_str` must be a string of
  length at least 1. Additionally, because `rest_as_str` is a substring
  of `n_as_str` and that is a string representation of a number,
  `rest_as_str` must be a string representation of a number as well.

Renaming of `rest_as_str` to `rest_str` was done to keep the line length
at at most 80 characters throughout the source code of this project.